Fixed a bug for maximum number of sequences.
Rank 0 wrote strept-mpich2.fasta (contiguous sequences in FASTA format) 
Rank 0 wrote strept-mpich2.ReceivedMessages.txt (MPI communication matrix) 
        %  & numberOfContigs & bases & meanSize  & n50  & max   & coverage   & misassembled & mismatches & indels
 Ray & 98 & 2014191 & 20552 & 36692 &  127906 &  0.9874 & 0 & 0 & 0 \\

+++ code/ArrayOfReads.cpp
@@ -83,7 +83,7 @@ int ArrayOfReads::size(){
 	return m_elements;
 }
 
-Read*ArrayOfReads::at(int i){
+Read*ArrayOfReads::at(uint64_t i){
 	#ifdef ASSERT
 	assert(m_maxSize!=0);
 	assert(m_numberOfChunks!=0);
@@ -111,7 +111,7 @@ Read*ArrayOfReads::at(int i){
 	return theRead;
 }
 
-Read*ArrayOfReads::operator[](int i){
+Read*ArrayOfReads::operator[](uint64_t i){
 	return at(i);
 }
